fix(db): screened_stocks.market_cap Integer → BigInteger
원 단위 시가총액 (최대 ~수조) 이 Postgres Integer(int32, ~21억) 초과해
DataError "value out of int32 range" 발생. BigInteger 로 변경.

Symptom: 스크리닝 INSERT 실패 ($7=58719375000 for 평화홀딩스)

- models.py: market_cap: BigInteger
- 0006_market_cap_bigint: ALTER COLUMN (idempotent)

+++ backend/app/db/models.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import String, Integer, Float, Boolean, DateTime, ForeignKey, Enum as SAEnum
+from sqlalchemy import String, Integer, BigInteger, Float, Boolean, DateTime, ForeignKey, Enum as SAEnum
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 import enum
 from app.db.database import Base
@@ -67,7 +67,7 @@ class ScreenedStock(Base):
     current_price: Mapped[int] = mapped_column(Integer)
     high_1y: Mapped[int] = mapped_column(Integer)            # 1년 고점
     low_1y: Mapped[int] = mapped_column(Integer)             # 1년 저점
-    market_cap: Mapped[int] = mapped_column(Integer)         # 시가총액 (원)
+    market_cap: Mapped[int] = mapped_column(BigInteger)      # 시가총액 (원) — 조 단위라 BigInteger
     net_income: Mapped[float] = mapped_column(Float)         # 순이익 (억원)
     operating_income: Mapped[float] = mapped_column(Float)   # 영업이익 (억원)
     foreign_ratio: Mapped[float] = mapped_column(Float)      # 외국인 비율 (%)
